+++ CHANGELOG.rst
@@ -12,7 +12,7 @@ Added
 
 Changed
 ~~~~~~~
-*
+* Replaced ``[test]`` and ``[docs]`` with a single ``[dev]`` installation extras target. Removed 'requirements.txt'. (#208)
 
 Deprecated
 ~~~~~~~~~~